Better fixes than code rabbit suggested

+++ src/synthesizer/extensions/reductions.cpp
@@ -66,7 +66,7 @@ static void reduce_spectra_parallel(double *spectra, double *part_spectra,
     // Thread-local accumulation to avoid false sharing and atomics
     std::vector<double> local(nlam, 0.0);
 #pragma omp for nowait schedule(static)
-    for (int p = 0; p < npart; p++) {
+    for (size_t p = 0; p < npart; p++) {
       for (int ilam = 0; ilam < nlam; ilam++) {
         size_t part_spec_ind = p * nlam + ilam;
         local[ilam] += part_spectra[part_spec_ind];
